+++ OrderFood.Application/Foods/FoodService.cs
@@ -16,25 +16,64 @@ namespace OrderFood.Application.Foods
 
         public async Task<int> CreateAsync(FoodCreateDto foodCreateDto)
         {
+            //var foodArg = new FoodArg
+            //{
+            //    Title = foodCreateDto.Title,
+            //    Price = foodCreateDto.Price,
+            //    VendorId = foodCreateDto.VendorId,
+            //};
+
+            //var food = new Food(Guid.NewGuid(), foodArg);
             var food = foodCreateDto.MapToArgument();
             return await _repository.CreateAsync(food);
         }
         public async Task<int> UpdateAsync(FoodUpdateDto foodUpdateDto)
         {
             var food = await _repository.GetAsNoTracking().FirstOrDefaultAsync(f => f.Id == foodUpdateDto.Id);
+            //var foodArg = new FoodArg
+            //{
+            //    Title = foodUpdateDto.Title,
+            //    Price = foodUpdateDto.Price,
+            //    CreateDate = foodUpdateDto.CreateDate,
+            //    VendorId = foodUpdateDto.VendorId,
+            //};
+
+            //food = new Food(foodUpdateDto.Id, foodArg);
             food = foodUpdateDto.MapToArgument();
             return await _repository.UpdateAsync(food);
         }
 
         public async Task<FoodGetDto?> GetByIdAsync(Guid id)
         {
-            var food = await _repository.GetAsNoTracking().FirstOrDefaultAsync(x => x.Id == id) ?? throw new Exception();
-            var foodGetDto = new FoodGetDto();
-            return foodGetDto.MapToArgumentForGetById(food);
+            var food = await _repository.GetAsNoTracking().FirstOrDefaultAsync(x => x.Id == id);
+            if (food != null)
+            {
+                //return new()
+                //{
+                //    Id = food.Id,
+                //    Title = food.Title,
+                //    Price = food.Price,
+                //    CreateDate = food.CreateDate,
+                //    VendorId = food.VendorId,
+                //};
+                var foodGetDto = new FoodGetDto();
+                return foodGetDto.MapToArgumentForGetById(food);
+            }
+
+            return null;
         }
         public async Task<List<FoodGetDto>> GetAsync(int skip, int take)
         {
             var foods = await _repository.GetAsNoTracking().Skip(skip).Take(take).ToListAsync();
+            //var result = foods.Select(f => new FoodGetDto
+            //{
+            //    Title = f.Title,
+            //    Price = f.Price,
+            //    CreateDate = f.CreateDate,
+            //    VendorId = f.VendorId
+            //}).ToList();
+
+            //return result;
             var foodGetDto = new FoodGetDto();
             var result = foodGetDto.MapToArgument(foods);
             return result;
@@ -42,16 +81,22 @@ namespace OrderFood.Application.Foods
 
         public async Task<int> RemoveAsync(Guid id)
         {
-            var food = await _repository.GetAsNoTracking().FirstOrDefaultAsync(f => f.Id == id) ?? throw new Exception();
-            var foodArg = new FoodArg
+            var food = await _repository.GetAsNoTracking().FirstOrDefaultAsync(f => f.Id == id);
+            if (food != null)
             {
-                Title = food.Title,
-                Price = food.Price,
-                CreateDate = food.CreateDate,
-                VendorId = food.VendorId,
-            };
-            food = new Food(id, foodArg);
-            return await _repository.RemoveAsync(food);
+                var foodArg = new FoodArg
+                {
+                    Title = food.Title,
+                    Price = food.Price,
+                    CreateDate = food.CreateDate,
+                    VendorId = food.VendorId,
+                };
+                food = new Food(id, foodArg);
+                return await _repository.RemoveAsync(food);
+            }
+
+            return 0;
+
         }
     }
 }
